+++ app/src/db/users.ts
@@ -8,7 +8,9 @@ import {
   UserInteraction,
 } from "@prisma/client"
 import { AggregatedType } from "eas-indexer/src/types"
+import { revalidatePath } from "next/cache"
 
+import { auth } from "@/auth"
 import { CONTRIBUTOR_ELIGIBLE_PROJECTS } from "@/lib/constants"
 import { EXTENDED_TAG_BY_ENTITY } from "@/lib/constants"
 import { ExtendedAggregatedType } from "@/lib/types"
@@ -699,4 +701,45 @@ export async function addTags(records: EntityRecords) {
       tags: [],
     })),
   ]
+}
+
+export async function makeUserAddressPrimary(address: string) {
+  const session = await auth()
+  if (!session?.user) {
+    throw new Error("Unauthorized")
+  }
+
+  const existingPrimary = await prisma.userAddress.findFirst({
+    where: {
+      primary: true,
+      userId: session.user.id,
+    },
+  })
+  if (existingPrimary) {
+    await prisma.userAddress.update({
+      where: {
+        address_userId: {
+          address: existingPrimary.address,
+          userId: session.user.id,
+        },
+      },
+      data: {
+        primary: false,
+      },
+    })
+  }
+
+  await prisma.userAddress.update({
+    where: {
+      address_userId: {
+        address,
+        userId: session.user.id,
+      },
+    },
+    data: {
+      primary: true,
+    },
+  })
+
+  revalidatePath("/profile/verified-addresses")
 }
